+++ server-spi-private/src/main/java/org/keycloak/events/EventBuilder.java
@@ -51,6 +51,7 @@ public class EventBuilder {
     private static final Logger log = Logger.getLogger(EventBuilder.class);
     public static final String AUTHN_AUTHORITY ="authnAuthority";
     public static final String VO_PERSON_ID ="voPersonID";
+    public static final String IDP_NAME = "idpName";
     
     private final KeycloakSession session;
     private EventStoreProvider store;
@@ -129,8 +130,6 @@ public class EventBuilder {
 
     public EventBuilder user(UserModel user) {
         event.setUserId(user == null ? null : user.getId());
-        if (user != null && user.getAttributeStream(AUTHN_AUTHORITY).count() >0)
-            detail(AUTHN_AUTHORITY,user.getAttributeStream(AUTHN_AUTHORITY).collect(Collectors.toList()));
         if (user != null && user.getAttributeStream(VO_PERSON_ID).count() >0)
             detail(VO_PERSON_ID,user.getAttributeStream(VO_PERSON_ID).collect(Collectors.toList()));
         return this;
@@ -143,6 +142,7 @@ public class EventBuilder {
 
     public EventBuilder session(UserSessionModel session) {
         event.setSessionId(session == null ? null : session.getId());
+        this.detail(AUTHN_AUTHORITY, session.getNote(Details.IDENTITY_PROVIDER_ID));
         return this;
     }