+++ services/src/main/java/org/keycloak/services/resources/LoginActionsService.java
@@ -55,6 +55,7 @@ import org.keycloak.events.Errors;
 import org.keycloak.events.EventBuilder;
 import org.keycloak.events.EventType;
 import org.keycloak.exceptions.TokenNotActiveException;
+import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakContext;
 import org.keycloak.models.SingleUseObjectKeyModel;
 import org.keycloak.models.AuthenticationFlowModel;
@@ -95,6 +96,7 @@ import org.keycloak.services.util.AuthenticationFlowURLHelper;
 import org.keycloak.services.util.BrowserHistoryHelper;
 import org.keycloak.services.util.CacheControlUtil;
 import org.keycloak.services.util.LocaleUtil;
+import org.keycloak.services.util.UserSessionUtil;
 import org.keycloak.sessions.AuthenticationSessionModel;
 import org.keycloak.sessions.RootAuthenticationSessionModel;
 
@@ -926,8 +928,7 @@ public class LoginActionsService {
         }
 
         event.detail(Details.IDENTITY_PROVIDER, identityProviderAlias)
-                .detail(Details.IDENTITY_PROVIDER_USERNAME, brokerContext.getUsername())
-                .detail(Details.IDENTITY_PROVIDER_BROKER_SESSION_ID, brokerContext.getBrokerSessionId());
+                .detail(Details.IDENTITY_PROVIDER_USERNAME, brokerContext.getUsername());
 
         event.success();
 
@@ -1111,8 +1112,14 @@ public class LoginActionsService {
         Map<String, String> userSessionNotes = authSession.getUserSessionNotes();
         String identityProvider = userSessionNotes.get(Details.IDENTITY_PROVIDER);
         if (identityProvider != null) {
+            IdentityProviderModel idp = authSession.getRealm().getIdentityProviderByAlias(identityProvider);
             event.detail(Details.IDENTITY_PROVIDER, identityProvider)
-                    .detail(Details.IDENTITY_PROVIDER_USERNAME, userSessionNotes.get(Details.IDENTITY_PROVIDER_USERNAME));
+                    .detail(Details.IDENTITY_PROVIDER_USERNAME, userSessionNotes.get(Details.IDENTITY_PROVIDER_USERNAME))
+                    .detail(EventBuilder.AUTHN_AUTHORITY, userSessionNotes.get(Details.IDENTITY_PROVIDER_ID))
+                    .detail(EventBuilder.IDP_NAME, UserSessionUtil.getIdPName(idp));
+            if (userSessionNotes.containsKey(Details.IDENTITY_PROVIDER_AUTHN_AUTHORITIES)){
+                this.event.detail(Details.IDENTITY_PROVIDER_AUTHN_AUTHORITIES, userSessionNotes.get(Details.IDENTITY_PROVIDER_AUTHN_AUTHORITIES));
+            }
         }
     }