+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 - Allow forwarding OIDC scopes to upstream OIDC Identity Provider in Keycloak [Allow forwarding OIDC scopes to upstream OIDC Identity Provider in Keycloak](https://trello.com/c/9I5SeGN6/2470-allow-forwarding-oidc-scopes-to-upstream-oidc-identity-provider-in-keycloak)
 - Add optional scopes for OIDC IdP and related changes
 - OIDC UserAttribute mapper strategy FORCE will not delete not existing values if pass scope is enabled and scope is not passed [Implement mapper strategy FORCE update only when claim is present](https://trello.com/c/ClFiAOgF/2543-implement-mapper-strategy-force-update-only-when-claim-is-present)
+- Handling authenticating authority(ies) in user session
 
 ### Changed
 - Change emailVerified User field with UserAttributeMappers (conditional trust email). [EOSC-KC-70](https://github.com/eosc-kc/keycloak/issues/70)